+++ playground/optimize-linked/linked-dep/index.js
@@ -1,0 +1,3 @@
+import { foo } from './other.js'
+
+export { foo }